Enhancement Two.
DHBandler.java: Added code to ensure that all cursors are closed appropriately, preventing cursor leaks. Added the fuzzy search method. Added a few helper methods to reduce and prevent duplicate code. Eliminated the SQLiteDatabase as an argument in a few methods because it's easily retrievable anywhere in this class.
SearchCampsitesActivity.java: This activity now uses the fuzzy search method to retrieve an ArrayList of results rather than a single exact match.